requirements.txt
praw
python-dotenv
psycopg2-binary
feedparser
websockets